requirements.txt
python-Levenshtein
nicknames
